+++ src/main/java/Backend/teampple/domain/auth/form/CustomUserDetailServiceImpl.java
@@ -1,4 +1,4 @@
-package Backend.teampple.domain.auth.security;
+package Backend.teampple.domain.auth.form;
 
 import Backend.teampple.domain.users.entity.User;
 import Backend.teampple.domain.users.repository.UserRepository;
@@ -18,6 +18,7 @@ import java.util.Collections;
 @Transactional(readOnly = true)
 /**
  * Spring Security 에서 사용자 정보를 담는 인터페이스
+ * Form 형식으로 구성 , Teampple Form 형식 미구현
  * */
 public class CustomUserDetailServiceImpl implements UserDetailsService {
     private final UserRepository userRepository;
@@ -27,7 +28,8 @@ public class CustomUserDetailServiceImpl implements UserDetailsService {
         User user = userRepository.findByKakaoId(username)
                 .orElseThrow(() -> new NotFoundException("[" + username + "] 사용자를 찾을 수 없습니다."));
         return CustomUserDetails.builder()
-                .kakaoId(user.getKakaoId())
+//                .id(user.getUserId())
+//                .password(user.getPassword)
                 .isDeleted(user.isDeleted())
                 .authorities(Collections.singleton(new SimpleGrantedAuthority("ROLE_USER")))
                 .build();
