+++ src/main/java/Backend/teampple/domain/auth/form/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package Backend.teampple.domain.auth.security;
+package Backend.teampple.domain.auth.form;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,15 +7,21 @@ import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.Collection;
 
+/**
+ * customUserDetail -> Authentication 내부에 들어있는 User 객체 값
+ * userName = userEntity 를 조회할 수 있는 uniqueKey
+ */
 @Getter
 public class CustomUserDetails implements UserDetails {
-    private String kakaoId;
-    private boolean isDeleted;
-    private Collection<GrantedAuthority> authorities;
+    private final String id;
+    private final String password;
+    private final boolean isDeleted;
+    private final Collection<GrantedAuthority> authorities;
 
     @Builder
-    public CustomUserDetails(String kakaoId, boolean isDeleted, Collection<GrantedAuthority> authorities) {
-        this.kakaoId = kakaoId;
+    public CustomUserDetails(String id, String password, boolean isDeleted, Collection<GrantedAuthority> authorities) {
+        this.id = id;
+        this.password = password;
         this.isDeleted = isDeleted;
         this.authorities = authorities;
     }
@@ -27,12 +33,12 @@ public class CustomUserDetails implements UserDetails {
 
     @Override
     public String getPassword() {
-        return "";
+        return password;
     }
 
     @Override
     public String getUsername() {
-        return kakaoId;
+        return id;
     }
 
     @Override
